+++ skforecast/ForecasterAutoregMultiVariate/tests/test_predict_quantiles.py
@@ -1,4 +1,4 @@
-# Unit test predict_dist ForecasterAutoregMultiVariate
+# Unit test predict_quantiles ForecasterAutoregMultiVariate
 # ==============================================================================
 import numpy as np
 import pandas as pd
@@ -7,7 +7,6 @@ from sklearn.linear_model import LinearRegression
 from sklearn.compose import ColumnTransformer
 from sklearn.preprocessing import StandardScaler
 from sklearn.preprocessing import OneHotEncoder
-from scipy.stats import norm
 
 # Fixtures
 from .fixtures_ForecasterAutoregMultiVariate import series
@@ -22,9 +21,9 @@ transformer_exog = ColumnTransformer(
                    )
 
 
-def test_predict_dist_output_when_forecaster_is_LinearRegression_steps_is_2_in_sample_residuals_True_exog_and_transformer():
+def test_predict_quantiles_output_when_forecaster_is_LinearRegression_steps_is_2_in_sample_residuals_True_exog_and_transformer():
     """
-    Test output of predict_dist when regressor is LinearRegression,
+    Test output of predict_quantiles when regressor is LinearRegression,
     2 steps are predicted, using in-sample residuals, exog is included and both
     inputs are transformed.
     """
@@ -38,27 +37,27 @@ def test_predict_dist_output_when_forecaster_is_LinearRegression_steps_is_2_in_s
                  )
     
     forecaster.fit(series=series, exog=exog)
-    results = forecaster.predict_dist(
+    results = forecaster.predict_quantiles(
                   steps               = 2,
                   exog                = exog_predict,
-                  distribution        = norm,
+                  quantiles           = [0.05, 0.55, 0.95],
                   n_boot              = 4,
                   in_sample_residuals = True
               )
     
     expected = pd.DataFrame(
-                   data    = np.array([[0.53808076, 0.11721631],
-                                       [0.48555937, 0.26296157]]),
-                   columns = ['loc', 'scale'],
+                   data    = np.array([[0.39855187, 0.56310088, 0.67329092],
+                                       [0.17729748, 0.48969677, 0.81780586]]),
+                   columns = ['q_0.05', 'q_0.55', 'q_0.95'],
                    index   = pd.RangeIndex(start=50, stop=52)
                )
     
     pd.testing.assert_frame_equal(expected, results)
 
 
-def test_predict_dist_output_when_forecaster_is_LinearRegression_steps_is_2_in_sample_residuals_False_exog_and_transformer():
+def test_predict_quantiles_output_when_forecaster_is_LinearRegression_steps_is_2_in_sample_residuals_False_exog_and_transformer():
     """
-    Test output of predict_dist when regressor is LinearRegression,
+    Test output of predict_quantiles when regressor is LinearRegression,
     2 steps are predicted, using out-sample residuals, exog is included and both
     inputs are transformed.
     """
@@ -73,18 +72,18 @@ def test_predict_dist_output_when_forecaster_is_LinearRegression_steps_is_2_in_s
     
     forecaster.fit(series=series, exog=exog)
     forecaster.out_sample_residuals = forecaster.in_sample_residuals
-    results = forecaster.predict_dist(
+    results = forecaster.predict_quantiles(
                   steps               = 2,
                   exog                = exog_predict,
-                  distribution        = norm,
+                  quantiles           = [0.05, 0.55, 0.95],
                   n_boot              = 4,
                   in_sample_residuals = False
               )
     
     expected = pd.DataFrame(
-                   data    = np.array([[0.53808076, 0.11721631],
-                                       [0.48555937, 0.26296157]]),
-                   columns = ['loc', 'scale'],
+                   data    = np.array([[0.39855187, 0.56310088, 0.67329092],
+                                       [0.17729748, 0.48969677, 0.81780586]]),
+                   columns = ['q_0.05', 'q_0.55', 'q_0.95'],
                    index   = pd.RangeIndex(start=50, stop=52)
                )